feat: Adicionando a verificação se o User professor tem cursos vinculados antes da exclusão

+++ src/services/userServices/DeleteUserService.ts
@@ -1,5 +1,5 @@
 import { IUserRepository } from '../../interfaces/IUserRepository'
-import { checkIfUserExists } from '../utils/checkFunctions'
+import { checkIfUserExists, checkUserLinkedToCourses } from '../utils/checkFunctions'
 import { CustomError } from '../../errors/CustomErrors'
 
 class DeleteUserService {
@@ -10,6 +10,12 @@ class DeleteUserService {
         const userExists = await checkIfUserExists(idUser)
         if(!userExists){
             return new CustomError('Usuário não encontrado.', 404)
+        }
+
+        //Verificando se o usuario esta vinculado a cursos caso ele seja professor
+        const isUserLinkedToCourses = await checkUserLinkedToCourses(idUser)
+        if(isUserLinkedToCourses){
+            return new CustomError('Este professor possui cursos cadastrados.', 422)
         }
 
         const deletedUser = await this.UserService.delete(idUser)
